Redesign of UI and update JS
I've done a lot of work I've been meaning to do for the last 3 years. I changed the layout a little bit, more to come on that, with a responsive UI design. Also added functionality for cycling through each game in the XML file. Still need to get the page to display the images for each game, for some reason the code is not working for that. Moved a lot of the functions out of the page, and into their own files, and eliminate the need for global variables by passing values from function to function. Huge progress!